+++ .lintstagedrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  '*.{ts,tsx,js,css,md}': 'prettier --write',
+};