+++ docs/src/components/Layout/Header.tsx
@@ -28,7 +28,7 @@ import LinkButton from './LinkButton';
 import { Logo } from '@/components/Logo';
 import NextLink from 'next/link';
 import { SecondaryNav } from './SecondaryNav';
-import { useRouter } from 'next/router';
+import { useCustomRouter } from '@/components/useCustomRouter';
 
 const NavLink = ({
   href,
@@ -41,8 +41,10 @@ const NavLink = ({
   isExternal?: boolean;
   onClick?: React.MouseEventHandler<HTMLAnchorElement>;
 }) => {
-  const { pathname, query } = useRouter();
-  const isCurrent = pathname.startsWith(href) && href !== '/';
+  const { pathname } = useCustomRouter();
+  const pathnameHeader = pathname.split('/')[2];
+  const hrefHeader = href.split('/')[2];
+  const isCurrent = pathnameHeader === hrefHeader && href !== `/`;
   const className = `docs-nav-link ${isCurrent ? 'current' : ''}`;
 
   if (isExternal) {
@@ -53,7 +55,7 @@ const NavLink = ({
     );
   }
   return (
-    <NextLink href={{ pathname: href, query }} passHref>
+    <NextLink href={href} passHref>
       <LinkButton href={href} classNames={className} onClick={onClick}>
         {children}
       </LinkButton>
@@ -62,17 +64,27 @@ const NavLink = ({
 };
 
 const Nav = (props) => (
-  <Flex as="nav" className="docs-nav" alignItems="center" gap="0" grow="1">
-    <NavLink {...props} href="/getting-started/installation">
+  <Flex
+    as="nav"
+    aria-label="Main navigation"
+    className="docs-nav"
+    alignItems="center"
+    gap="0"
+    grow="1"
+  >
+    <NavLink
+      {...props}
+      href={`/${props.platform}/getting-started/installation`}
+    >
       Getting started
     </NavLink>
-    <NavLink {...props} href="/components">
+    <NavLink {...props} href={`/${props.platform}/components`}>
       Components
     </NavLink>
-    <NavLink {...props} href="/theming">
+    <NavLink {...props} href={`/${props.platform}/theming`}>
       Theming
     </NavLink>
-    <NavLink {...props} href="/guides">
+    <NavLink {...props} href={`/${props.platform}/guides`}>
       Guides
     </NavLink>
     <Divider orientation="vertical" />
@@ -104,15 +116,15 @@ const ColorModeSwitcher = ({ colorMode, setColorMode }) => {
       isSelectionRequired
       className="color-switcher"
     >
-      <ToggleButton value="light">
+      <ToggleButton value="light" title="Light mode">
         <VisuallyHidden>Light mode</VisuallyHidden>
         <MdWbSunny />
       </ToggleButton>
-      <ToggleButton value="dark">
+      <ToggleButton value="dark" title="Dark mode">
         <VisuallyHidden>Dark mode</VisuallyHidden>
         <MdBedtime />
       </ToggleButton>
-      <ToggleButton value="system">
+      <ToggleButton value="system" title="System preferences">
         <VisuallyHidden>System preference</VisuallyHidden>
         <MdTonality />
       </ToggleButton>
@@ -138,14 +150,15 @@ export const Header = ({ platform, colorMode, setColorMode }) => {
           )}
         </Button>
 
-        <NavLink href="/">
+        <NavLink href={`/`}>
           <span className="docs-logo-link">
             <VisuallyHidden>Amplify UI Home</VisuallyHidden>
             <Logo />
           </span>
         </NavLink>
 
-        <Nav />
+        <Nav platform={platform} />
+
         <Settings
           colorMode={colorMode}
           setColorMode={setColorMode}
@@ -165,8 +178,8 @@ export const Header = ({ platform, colorMode, setColorMode }) => {
             />
           </Flex>
 
-          <Nav onClick={() => setExpanded(false)} />
-          <nav className="docs-sidebar-nav">
+          <Nav onClick={() => setExpanded(false)} platform={platform} />
+          <nav aria-label="Section navigation" className="docs-sidebar-nav">
             <SecondaryNav onClick={() => setExpanded(false)} />
           </nav>
         </View>