+++ docs/src/pages/[platform]/components/radiogroupfield/demo.tsx
@@ -1,0 +1,71 @@
+import * as React from 'react';
+import {
+  Radio,
+  RadioGroupField,
+  RadioGroupFieldProps,
+} from '@aws-amplify/ui-react';
+
+import { Demo } from '@/components/Demo';
+import { RadioGroupFieldPropControls } from './RadioGroupFieldPropControls';
+import { useRadioGroupFieldProps } from './useRadioGroupFieldProps';
+import { demoState } from '@/utils/demoState';
+
+const propsToCode = ({
+  label,
+  labelPosition,
+  size,
+  direction,
+  name,
+  isDisabled,
+}: RadioGroupFieldProps) => {
+  return (
+    `<RadioGroupField` +
+    (label ? `\n  label=${JSON.stringify(label)}` : '') +
+    (name ? `\n  name=${JSON.stringify(name)}` : '') +
+    (isDisabled ? `\n  isDisabled={${isDisabled}}` : '') +
+    (labelPosition
+      ? `\n  labelPosition=${JSON.stringify(labelPosition)}`
+      : '') +
+    (direction ? `\n  direction=${JSON.stringify(direction)}` : '') +
+    (size ? `\n  size=${JSON.stringify(size)}` : '') +
+    `\n>` +
+    `\n  <Radio value="html">html</Radio>` +
+    `\n  <Radio value="css">css</Radio>` +
+    `\n  <Radio value="javascript">javascript</Radio>` +
+    `\n</RadioGroupField>`
+  );
+};
+
+const defaultRadioGroupFieldProps = {
+  label: 'Language',
+  name: 'language',
+  defaultValue: 'html',
+};
+
+export const RadioGroupFieldDemo = () => {
+  const props = useRadioGroupFieldProps(
+    (demoState.get(RadioGroupField.displayName) as RadioGroupFieldProps) ||
+      defaultRadioGroupFieldProps
+  );
+
+  return (
+    <Demo
+      code={propsToCode(props)}
+      propControls={<RadioGroupFieldPropControls {...props} />}
+    >
+      <RadioGroupField
+        name={props.name}
+        label={props.label}
+        defaultValue={props.defaultValue}
+        labelPosition={props.labelPosition}
+        isDisabled={props.isDisabled}
+        direction={props.direction}
+        size={props.size}
+      >
+        <Radio value="html">html</Radio>
+        <Radio value="css">css</Radio>
+        <Radio value="javascript">javascript</Radio>
+      </RadioGroupField>
+    </Demo>
+  );
+};