+++ docs/src/theme.ts
@@ -1,4 +1,4 @@
-import { defaultTheme, Theme } from '@aws-amplify/ui-react';
+import { defaultDarkModeOverride, Theme } from '@aws-amplify/ui-react';
 
 const flipper = {
   100: 10,
@@ -10,16 +10,6 @@ const flipper = {
   10: 100,
 };
 
-export const flipPalette = (obj) => {
-  return Object.keys(obj).reduce((acc, curr) => {
-    const { value } = obj[curr];
-    return {
-      ...acc,
-      [flipper[curr]]: { value },
-    };
-  }, {});
-};
-
 const usePalette = (str) => {
   return Object.keys(flipper).reduce((acc, curr) => {
     return {
@@ -29,45 +19,15 @@ const usePalette = (str) => {
   }, {});
 };
 
+export const baseTheme: Theme = {
+  name: 'amplify-docs',
+  overrides: [defaultDarkModeOverride],
+};
+
 export const theme: Theme = {
   name: 'amplify-docs',
   overrides: [
-    {
-      colorMode: 'dark',
-      tokens: {
-        colors: {
-          red: flipPalette(defaultTheme.tokens.colors.red),
-          orange: flipPalette(defaultTheme.tokens.colors.orange),
-          yellow: flipPalette(defaultTheme.tokens.colors.yellow),
-          green: flipPalette(defaultTheme.tokens.colors.green),
-          teal: flipPalette(defaultTheme.tokens.colors.teal),
-          blue: flipPalette(defaultTheme.tokens.colors.blue),
-          purple: flipPalette(defaultTheme.tokens.colors.purple),
-          pink: flipPalette(defaultTheme.tokens.colors.pink),
-          neutral: flipPalette(defaultTheme.tokens.colors.neutral),
-          black: { value: '#fff' },
-          white: { value: '#000' },
-
-          border: {
-            primary: { value: '{colors.neutral.20.value}' },
-            secondary: { value: '{colors.neutral.20.value}' },
-            tertiary: { value: '{colors.neutral.20.value}' },
-          },
-
-          overlay: {
-            10: { value: 'hsla(0, 0%, 100%, 0.1)' },
-            20: { value: 'hsla(0, 0%, 100%, 0.2)' },
-            30: { value: 'hsla(0, 0%, 100%, 0.3)' },
-            40: { value: 'hsla(0, 0%, 100%, 0.4)' },
-            50: { value: 'hsla(0, 0%, 100%, 0.5)' },
-            60: { value: 'hsla(0, 0%, 100%, 0.6)' },
-            70: { value: 'hsla(0, 0%, 100%, 0.7)' },
-            80: { value: 'hsla(0, 0%, 100%, 0.8)' },
-            90: { value: 'hsla(0, 0%, 100%, 0.9)' },
-          },
-        },
-      },
-    },
+    defaultDarkModeOverride,
     {
       selector:
         '[data-amplify-theme-override="classic"] [data-amplify-theme="amplify-docs"]',