+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/confirm-reset-password/amplify-confirm-reset-password.component.ts
@@ -3,6 +3,7 @@ import { AuthenticatorService } from '../../../../services/authenticator.service
 import {
   FormFieldsArray,
   getFormDataFromEvent,
+  hasTranslation,
   translate,
 } from '@aws-amplify/ui';
 
@@ -15,10 +16,17 @@ export class ConfirmResetPasswordComponent {
   @Input() public headerText = translate('Reset your password');
 
   // translated strings
-  public sendCodeText = translate('Send Code');
   public backToSignInText = translate('Back to Sign In');
   public resendCodeText = translate('Resend Code');
   public sortedFormFields: FormFieldsArray;
+  /**
+   * Support backwards compatibility for erroneous 'Send Code' text
+   * See https://github.com/aws-amplify/amplify-ui/issues/1784
+   * TODO: Remove support for 'Send Code' translation in next Major release
+   */
+  public submitText = !hasTranslation('Send Code')
+    ? translate('Submit')
+    : translate('Send Code');
 
   constructor(public authenticator: AuthenticatorService) {}