+++ packages/react/src/primitives/Alert/Alert.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import classNames from 'classnames';
 
 import { ComponentClassNames } from '../shared/constants';
+import { classNameModifier } from '../shared/utils';
 import { AlertProps, Primitive } from '../types';
 import { View } from '../View';
 import { Flex } from '../Flex';
@@ -37,7 +38,11 @@ const AlertPrimitive: Primitive<AlertProps, typeof Flex> = (
   return (
     !dismissed && (
       <Flex
-        className={classNames(ComponentClassNames.Alert, className)}
+        className={classNames(
+          ComponentClassNames.Alert,
+          className,
+          classNameModifier(ComponentClassNames.Alert, variation)
+        )}
         data-variation={variation}
         ref={ref}
         {...rest}