+++ packages/ui/src/machines/authenticator/index.ts
@@ -1,4 +1,5 @@
 import { assign, createMachine, forwardTo, spawn } from 'xstate';
+import { choose } from 'xstate/lib/actions';
 
 import {
   AuthContext,
@@ -178,6 +179,24 @@ export function createAuthenticatorMachine() {
           },
         },
         authenticated: {
+          initial: 'idle',
+          states: {
+            idle: {
+              on: {
+                TOKEN_REFRESH: 'refreshUser',
+              },
+            },
+            refreshUser: {
+              invoke: {
+                src: 'getCurrentUser',
+                onDone: {
+                  actions: 'setUser',
+                  target: 'idle',
+                },
+                onError: { target: '#authenticator.signOut' },
+              },
+            },
+          },
           on: { SIGN_OUT: 'signOut' },
         },
       },
@@ -187,14 +206,18 @@ export function createAuthenticatorMachine() {
         SUBMIT: { actions: 'forwardToActor' },
         FEDERATED_SIGN_IN: { actions: 'forwardToActor' },
         RESEND: { actions: 'forwardToActor' },
-        SIGN_OUT: { actions: 'forwardToActor' },
         SIGN_IN: { actions: 'forwardToActor' },
         SKIP: { actions: 'forwardToActor' },
       },
     },
     {
       actions: {
-        forwardToActor: forwardTo((context) => context.actorRef),
+        forwardToActor: choose([
+          {
+            cond: 'hasActor',
+            actions: forwardTo((context) => context.actorRef),
+          },
+        ]),
         setUser: assign({
           user: (_, event) => event.data as CognitoUserAmplify,
         }),
@@ -351,6 +374,8 @@ export function createAuthenticatorMachine() {
         shouldRedirectToResetPassword: (_, event) =>
           event.data?.intent === 'confirmPasswordReset',
         shouldSetup: (context) => context.hasSetup === false,
+        // other context guards
+        hasActor: (context) => !!context.actorRef,
       },
       services: {
         getCurrentUser: (context, _) => context.services.getCurrentUser(),