+++ packages/vue/src/components/sign-up.vue
@@ -23,7 +23,7 @@ const createAccountLabel = computed(() => translate('Create Account'));
 // Methods
 
 const onInput = (e: Event): void => {
-  let { checked, name, type, value } = <HTMLInputElement>e.target;
+  let { checked, name, type, value } = e.target as HTMLInputElement;
 
   if (type === 'checkbox' && !checked)
     (value as string | undefined) = undefined;
@@ -31,7 +31,7 @@ const onInput = (e: Event): void => {
 };
 
 function onBlur(e: Event) {
-  const { name } = <HTMLInputElement>e.target;
+  const { name } = e.target as HTMLInputElement;
   props.updateBlur({ name });
 }
 
@@ -60,10 +60,9 @@ const submit = (e: Event): void => {
       >
         <federated-sign-in></federated-sign-in>
 
-        <base-wrapper class="amplify-flex" style="flex-direction: column">
+        <base-wrapper class="amplify-flex amplify-authenticator__column">
           <base-field-set
-            class="amplify-flex"
-            style="flex-direction: column"
+            class="amplify-flex amplify-authenticator__column"
             :disabled="isPending"
           >
             <template #fieldSetI="{ slotData }">
@@ -75,10 +74,10 @@ const submit = (e: Event): void => {
             {{ translate(error) }}
           </base-alert>
           <amplify-button
-            class="amplify-field-group__control"
-            data-fullwidth="true"
-            data-loading="false"
-            data-variation="primary"
+            class="amplify-field-group__control amplify-authenticator__font"
+            :fullwidth="true"
+            :loading="false"
+            :variation="'primary'"
             style="border-radius: 0px; font-weight: normal"
             :disabled="isPending || hasValidationErrors"
             >{{ createAccountLabel }}</amplify-button